Okay, When we try implment Octree suddenly we find out we need the the bounding box of entity, so we turend to finding out entity bounding volume and rendering is on demand, and we just did that, in the mesh component under the bounding volume we can tick render Volume and the bounding volume will be shown, for kinned and unskninned mesh, which was awesome, now we off to the octree, yeah one last note when we clone entity the flag is not copyed along with other stuff we fix it when the time arives for now it is DONE, cool

+++ Core/Core/Component/SkinnedMeshComponent/SkinnedMeshComponent.cpp
@@ -54,7 +54,7 @@ namespace OE1Core
 		void SkinnedMeshComponent::SetOffset(GLintptr _offset)
 		{
 			m_Offset = _offset;
-			m_CoreInstanceData.AnimationID = _offset;
+			m_CoreInstanceData.AnimationID = (int)_offset;
 		}
 		GLintptr SkinnedMeshComponent::GetOffset() const
 		{
